requirements.txt
beautifulsoup4
requests
selenium
lxml
fake_useragent
undetected-chromedriver
pyyaml